+++ src/app/(common)/mypage/_components/ReviewModal/FormActions.tsx
@@ -1,0 +1,19 @@
+import Button from "@/components/Button";
+
+type FormActionsProps = {
+  onCancel: () => void;
+  isSubmitDisabled: boolean;
+};
+
+export const FormActions: React.FC<FormActionsProps> = ({ onCancel, isSubmitDisabled }) => {
+  return (
+    <div className="flex gap-4">
+      <Button className="w-full" styleType={"outline"} onClick={onCancel}>
+        취소
+      </Button>
+      <Button className="w-full" type="submit" disabled={isSubmitDisabled}>
+        리뷰 등록
+      </Button>
+    </div>
+  );
+};
