+++ src/app/(common)/mypage/_components/ReviewModal/index.ts
@@ -1,0 +1,1 @@
+export { default as ReviewModal } from "./ReviewModal";
